Phase 102.5: Resolve typedef names to struct names in type tracking
When storing type information for variables and parameters, now resolve
typedef'd struct names to their underlying struct names. This allows member
access (->  and .) to find struct definitions correctly when using typedef'd
pointers.

Changes:
- Resolve typedefs in localTypeNames_ for function parameters
- Resolve typedefs in localTypeNames_ for global variables
- Resolve typedefs in localTypeNames_ for local variables
- Add typedef resolution in getExprTypeInfo for member access validation
- Pre-pass registerAllStructDefinitions ensures structs are available

Fixes compiler errors when accessing struct members through typedef'd pointers,
e.g., typedef struct Point { int x; } Point_t;
       void func(Point_t *p) { p->x = 10; }  // Now works correctly

+++ lib/stdlib/memcpy_fast.c
@@ -51,8 +51,8 @@ void *memcpy_optimized(void *dest, const void *src, int n) {
     }
 
     /* Detect alignment: if both pointers aligned to 4 bytes, use fast path */
-    unsigned long d_addr = (unsigned long)(uintptr_t)dest;
-    unsigned long s_addr = (unsigned long)(uintptr_t)src;
+    unsigned long d_addr = (unsigned long)dest;
+    unsigned long s_addr = (unsigned long)src;
 
     if ((d_addr & 3) == 0 && (s_addr & 3) == 0) {
         return memcpy_aligned_4((char *)dest, (const char *)src, n);
